+++ src/main/java/com/tmxk/map/entity/ProWorkEntity.java
@@ -23,7 +23,7 @@ public class ProWorkEntity implements Serializable {
 	 * 人才id
 	 */
 	@TableId
-	private Integer professionalId;
+	private Long professionalId;
 	/**
 	 * 公司名
 	 */
